simplify AppState
